Remove utils/search.py
