+++ flexget/plugins/input/gazelle.py
@@ -1,8 +1,10 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
-from datetime import datetime
 import logging
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
+from datetime import datetime
+
+from sqlalchemy import Column, Unicode, String, DateTime
 
 from flexget import plugin, db_schema
 from flexget.config_schema import one_or_more
@@ -12,11 +14,7 @@ from flexget.manager import Session
 from flexget.plugin import PluginError
 from flexget.utils.database import json_synonym
 from flexget.utils.requests import TokenBucketLimiter
-from flexget.utils.search import normalize_unicode
 from flexget.utils.tools import parse_filesize
-
-from sqlalchemy import Column, Unicode, String, DateTime
-
 
 DETECT_2FA = ("Authenticator Code", "TOTP code")
 log = logging.getLogger('gazelle')
