+++ flexget/plugins/sites/allyoulike.py
@@ -8,7 +8,7 @@ from requests.exceptions import RequestException
 
 from flexget import plugin
 from flexget.event import event
-from flexget.utils.search import normalize_unicode
+from flexget.plugins.sites import normalize_unicode
 from flexget.utils.soup import get_soup
 try:
     from flexget.plugins.internal.urlrewriting import UrlRewritingError
