+++ flexget/plugins/sites/cpasbien.py
@@ -10,7 +10,7 @@ from flexget.entry import Entry
 from flexget.event import event
 from flexget.utils import requests
 from flexget.utils.soup import get_soup
-from flexget.utils.search import normalize_unicode
+from flexget.plugins.sites import normalize_unicode
 from flexget.utils.tools import parse_filesize
 
 log = logging.getLogger('search_cpasbien')
