+++ flexget/plugins/sites/fuzer.py
@@ -12,7 +12,7 @@ from flexget.entry import Entry
 from flexget.event import event
 from flexget.plugin import PluginError
 from flexget.utils.requests import Session as RequestSession
-from flexget.utils.search import torrent_availability, normalize_scene
+from flexget.plugins.sites import normalize_scene, torrent_availability
 from flexget.utils.soup import get_soup
 from flexget.utils.tools import parse_filesize
