+++ flexget/plugins/sites/newtorrents.py
@@ -11,7 +11,7 @@ from flexget.entry import Entry
 from flexget.event import event
 from flexget.plugins.internal.urlrewriting import UrlRewritingError
 from flexget.utils.soup import get_soup
-from flexget.utils.search import torrent_availability, normalize_unicode
+from flexget.plugins.sites import normalize_unicode, torrent_availability
 from flexget.utils import requests
 
 timeout = 10
