+++ flexget/plugins/sites/ptn.py
@@ -10,7 +10,7 @@ from flexget.event import event
 from flexget.utils import requests
 from flexget.utils.imdb import extract_id
 from flexget.utils.soup import get_soup
-from flexget.utils.search import torrent_availability
+from flexget.plugins.sites import torrent_availability
 from flexget.utils.tools import parse_filesize
 
 log = logging.getLogger('search_ptn')
