+++ flexget/plugins/sites/site_1337x.py
@@ -10,7 +10,7 @@ from flexget.entry import Entry
 from flexget.event import event
 from flexget.utils.requests import RequestException
 from flexget.utils.soup import get_soup
-from flexget.utils.search import torrent_availability
+from flexget.plugins.sites import torrent_availability
 from flexget.utils.tools import parse_filesize
 from flexget.plugins.internal.urlrewriting import UrlRewritingError
