+++ flexget/plugins/sites/torrentleech.py
@@ -12,7 +12,7 @@ from flexget.entry import Entry
 from flexget.event import event
 from flexget.plugin import PluginError
 from flexget.plugins.internal.urlrewriting import UrlRewritingError
-from flexget.utils.search import torrent_availability, normalize_unicode
+from flexget.plugins.sites import normalize_unicode, torrent_availability
 from flexget.utils.tools import parse_filesize
 
 log = logging.getLogger('torrentleech')
