+++ flexget/plugins/sites/torrentz.py
@@ -12,7 +12,7 @@ import requests
 from flexget import plugin
 from flexget.entry import Entry
 from flexget.event import event
-from flexget.utils.search import torrent_availability, normalize_unicode
+from flexget.plugins.sites import normalize_unicode, torrent_availability
 
 log = logging.getLogger('torrentz')
